Add CMDSTAN_HOME and accessor.
Fixes problems with inferring CMDSTAN_HOME and using it as a default
argument by introducing it into the scope of the module. Since this
cannot be modifed outside the module, also add an accessor. Add
__init__() which infers a value from Main or ENV when the module is
first loaded.

Cosmetic changes: warn instead of println, remove indentation inside
module (as recommended in Julia).